requirements.txt
torch==2.7.0
torchaudio==2.7.0
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
safetensors==0.5.3
pyyaml==6.0.2
numpy==2.2.5
starlette==0.46.2
click==8.2.0
pydantic==2.11.4 